Added support for auto light sleep mode
Automatic switching to light sleep (or modem sleep) can be now
configured on startup so it can be used together with ESP-IDF's
power management component. On the other hand, a callback is entered
so user-specific actions can be executed right before entering
light sleep mode (for example, enabling GPIO wakeup).